# application/src/main/openapi/ledgerflow.yaml
openapi: 3.1.0
info:
  title: LedgerFlow API
  version: 0.4.0
  description: >-
    LedgerFlow's contract-first HTTP API. Creating an INR order synchronously drives the mock
    payment authorization/capture and durable financial finalization. Outbox publication and
    notification processing are asynchronous and are not prerequisites for the HTTP result.
servers:
  - url: http://localhost:8080
    description: Local application
tags:
  - name: Orders
paths:
  /api/v1/orders:
    post:
      tags: [Orders]
      operationId: createOrder
      summary: Create an order idempotently
      description: >-
        Creates and advances one INR order for the authenticated subject. Provider authorization
        and capture are synchronous, while outbox publication and notification processing remain
        asynchronous. Reusing the same Idempotency-Key with the same canonical request returns the
        original business result; reusing it with a different request returns 409. An unknown
        provider outcome is looked up using the original operation ID before any same-ID resend.
        The JWT must have the ledgerflow.orders.write scope and customer or admin realm role. Query
        parameters, compressed bodies, duplicate/unknown JSON properties, and request bodies larger
        than 16 KiB are rejected.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
            examples:
              standardOrder:
                summary: INR order
                value:
                  clientReference: checkout-20260713-0001
                  amount:
                    amountMinor: 259900
                    currency: INR
                  paymentMethodReference: pm_mock_success
      responses:
        "201":
          description: >-
            Original or replayed terminal order result. COMPLETED means payment CAPTURED and the
            balanced journal plus durable outbox event committed; it does not mean Kafka publication
            or notification processing completed. PAYMENT_DECLINED is also a terminal created order.
          headers:
            Location:
              $ref: "#/components/headers/Location"
            X-Correlation-Id:
              $ref: "#/components/headers/CorrelationId"
            Idempotency-Replayed:
              $ref: "#/components/headers/IdempotencyReplayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
              examples:
                completed:
                  $ref: "#/components/examples/CompletedOrder"
                declined:
                  $ref: "#/components/examples/DeclinedOrder"
        "202":
          description: >-
            The order and stable provider identities are durable, but a bounded temporary or unknown
            provider outcome requires a later retry with the same Idempotency-Key.
          headers:
            Location:
              $ref: "#/components/headers/Location"
            X-Correlation-Id:
              $ref: "#/components/headers/CorrelationId"
            Idempotency-Replayed:
              $ref: "#/components/headers/IdempotencyReplayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
              examples:
                retryPending:
                  $ref: "#/components/examples/RetryPendingOrder"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "406":
          $ref: "#/components/responses/NotAcceptable"
        "409":
          description: The idempotency key was already used for a different canonical request
          headers:
            X-Correlation-Id:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                keyReused:
                  value:
                    type: https://ledgerflow.example/problems/idempotency-key-reused
                    title: Idempotency key reused
                    status: 409
                    detail: The idempotency key was already used with a different request.
                    instance: /api/v1/orders
                    code: idempotency_key_reused
                    correlationId: order-conflict-001
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          description: The request uses a well-formed currency that this MVP does not support
          headers:
            X-Correlation-Id:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unsupportedCurrency:
                  value:
                    type: https://ledgerflow.example/problems/unsupported-currency
                    title: Unsupported currency
                    status: 422
                    detail: The MVP supports INR only.
                    instance: /api/v1/orders
                    code: unsupported_currency
                    correlationId: order-currency-001
        "500":
          $ref: "#/components/responses/InternalError"
        "502":
          description: >-
            The provider returned malformed or contradictory data. The failed order remains
            owner-readable at Location, and an identical idempotent replay returns this outcome.
          headers:
            Location:
              $ref: "#/components/headers/Location"
            X-Correlation-Id:
              $ref: "#/components/headers/CorrelationId"
            Idempotency-Replayed:
              $ref: "#/components/headers/IdempotencyReplayed"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                providerProtocol:
                  value:
                    type: https://ledgerflow.example/problems/provider-protocol-error
                    title: Payment provider protocol error
                    status: 502
                    detail: The payment provider returned an invalid response.
                    instance: /api/v1/orders
                    code: provider_protocol_error
                    correlationId: order-provider-001
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/orders/{orderId}:
    get:
      tags: [Orders]
      operationId: getOrder
      summary: Get the current owned order
      description: >-
        Requires the ledgerflow.orders.read scope and customer or admin realm role. Missing orders
        and orders owned by another JWT subject both return the same 404 response.
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: 019535d9-3df7-79fb-b466-fa907fa17f9e
        - $ref: "#/components/parameters/CorrelationId"
      responses:
        "200":
          description: Current order representation
          headers:
            X-Correlation-Id:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
              examples:
                current:
                  $ref: "#/components/examples/CompletedOrder"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "406":
          $ref: "#/components/responses/NotAcceptable"
        "404":
          description: Order absent or not owned by the authenticated subject
          headers:
            X-Correlation-Id:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                notFound:
                  value:
                    type: https://ledgerflow.example/problems/resource-not-found
                    title: Resource not found
                    status: 404
                    detail: The requested order was not found.
                    instance: /api/v1/orders/019535d9-3df7-79fb-b466-fa907fa17f9e
                    code: resource_not_found
                    correlationId: order-read-001
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        OAuth 2.0/OIDC JWT with the configured exact issuer and ledgerflow-api audience, signed with
        RS256. POST requires ledgerflow.orders.write; GET requires ledgerflow.orders.read. Active
        order routes also require the Keycloak customer or admin realm role.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        Case-sensitive client nonce. Validated before hashing; raw values are never persisted or
        logged.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: "^[A-Za-z0-9._:-]+$"
      example: order-019535d9-3df7-79fb
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: Invalid or absent values are replaced by the server.
      schema:
        type: string
        minLength: 1
        maxLength: 64
        pattern: "^[A-Za-z0-9._-]+$"
      example: checkout-order-001
  headers:
    Location:
      description: Canonical path of the created order
      schema:
        type: string
        example: /api/v1/orders/019535d9-3df7-79fb-b466-fa907fa17f9e
    CorrelationId:
      description: Valid correlation ID used for this HTTP operation
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 64
        pattern: "^[A-Za-z0-9._-]+$"
    IdempotencyReplayed:
      description: Present and true only when the original result is replayed
      required: false
      schema:
        type: boolean
  schemas:
    CreateOrderRequest:
      type: object
      additionalProperties: false
      required: [amount, paymentMethodReference]
      properties:
        clientReference:
          type: string
          minLength: 1
          maxLength: 100
          pattern: "^\\S(?:.*\\S)?$"
          example: checkout-20260713-0001
        amount:
          $ref: "#/components/schemas/Money"
        paymentMethodReference:
          type: string
          minLength: 9
          maxLength: 128
          pattern: "^pm_mock_[a-z_]+$"
          description: >-
            Opaque local/test mock-provider token. It is included in idempotency hashing, is never
            returned or logged, and must never contain card data or real payment credentials.
          example: pm_mock_success
    Money:
      type: object
      additionalProperties: false
      required: [amountMinor, currency]
      properties:
        amountMinor:
          type: integer
          format: int64
          minimum: 1
          example: 259900
        currency:
          type: string
          minLength: 3
          maxLength: 3
          pattern: "^[A-Z]{3}$"
          enum: [INR]
    Order:
      type: object
      additionalProperties: false
      required: [orderId, status, amount, payment, createdAt, updatedAt]
      properties:
        orderId:
          type: string
          format: uuid
          description: PostgreSQL 18 UUIDv7 identifier
          example: 019535d9-3df7-79fb-b466-fa907fa17f9e
        clientReference:
          type:
            - string
            - "null"
          minLength: 1
          maxLength: 100
        status:
          type: string
          enum:
            - CREATED
            - PAYMENT_PROCESSING
            - COMPLETED
            - PAYMENT_DECLINED
            - PAYMENT_RETRY_PENDING
            - FAILED
        amount:
          $ref: "#/components/schemas/Money"
        payment:
          description: >-
            Payment workflow state for Milestone 6 orders. Null is retained only so historical
            pre-Milestone-6 CREATED orders remain readable after an in-place database upgrade.
          oneOf:
            - $ref: "#/components/schemas/PaymentSummary"
            - type: "null"
        createdAt:
          type: string
          format: date-time
          example: 2026-07-13T09:42:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2026-07-13T09:42:01Z
    PaymentSummary:
      type: object
      additionalProperties: false
      required: [paymentId, status]
      properties:
        paymentId:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - CREATED
            - AUTHORIZING
            - AUTHORIZED
            - DECLINED
            - AUTHORIZATION_RETRY_PENDING
            - AUTHORIZATION_UNKNOWN
            - CAPTURING
            - CAPTURE_CONFIRMED
            - CAPTURE_ACCOUNTED
            - CAPTURED
            - CAPTURE_DECLINED
            - CAPTURE_RETRY_PENDING
            - CAPTURE_UNKNOWN
            - FAILED
        failureCode:
          type:
            - string
            - "null"
          maxLength: 64
          pattern: "^[A-Z0-9_]+$"
          example: AUTHORIZATION_DECLINED
    ValidationError:
      type: object
      additionalProperties: false
      required: [field, code]
      properties:
        field:
          type: string
          example: /amount/amountMinor
        code:
          type: string
          example: must_be_positive
    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status, detail, instance, code, correlationId]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
          pattern: "^[a-z0-9_]+$"
        correlationId:
          type: string
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"
  responses:
    InvalidRequest:
      description: Request body, path, or required header is invalid
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          examples:
            invalidRequest:
              value:
                type: https://ledgerflow.example/problems/invalid-request
                title: Invalid request
                status: 400
                detail: The request is invalid.
                instance: /api/v1/orders
                code: invalid_request
                correlationId: generated-correlation-id
                errors:
                  - field: /headers/Idempotency-Key
                    code: invalid_format
    Unauthorized:
      description: Missing or invalid bearer token
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Authenticated principal lacks the required scope or realm role
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotAcceptable:
      description: The requested response media type is not supported
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    PayloadTooLarge:
      description: The JSON request exceeds the configured 16 KiB limit
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    UnsupportedMediaType:
      description: The request is not uncompressed application/json
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    RateLimited:
      description: The authenticated subject exceeded the configured per-instance write limit
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
        Retry-After:
          description: Minimum whole seconds before this instance will admit another write
          required: true
          schema:
            type: integer
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    InternalError:
      description: Unexpected internal failure with no exposed internals
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ServiceUnavailable:
      description: Required durable infrastructure is unavailable
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  examples:
    CompletedOrder:
      value:
        orderId: 019535d9-3df7-79fb-b466-fa907fa17f9e
        clientReference: checkout-20260713-0001
        status: COMPLETED
        amount:
          amountMinor: 259900
          currency: INR
        payment:
          paymentId: 019535d9-4a20-7e29-9d35-a6e4aeab9e07
          status: CAPTURED
          failureCode: null
        createdAt: 2026-07-13T09:42:00Z
        updatedAt: 2026-07-13T09:42:01Z
    DeclinedOrder:
      value:
        orderId: 019535d9-3df7-79fb-b466-fa907fa17f9e
        clientReference: checkout-20260713-0002
        status: PAYMENT_DECLINED
        amount:
          amountMinor: 259900
          currency: INR
        payment:
          paymentId: 019535d9-4a20-7e29-9d35-a6e4aeab9e08
          status: DECLINED
          failureCode: AUTHORIZATION_DECLINED
        createdAt: 2026-07-13T09:42:00Z
        updatedAt: 2026-07-13T09:42:01Z
    RetryPendingOrder:
      value:
        orderId: 019535d9-3df7-79fb-b466-fa907fa17f9e
        clientReference: checkout-20260713-0003
        status: PAYMENT_RETRY_PENDING
        amount:
          amountMinor: 259900
          currency: INR
        payment:
          paymentId: 019535d9-4a20-7e29-9d35-a6e4aeab9e09
          status: AUTHORIZATION_RETRY_PENDING
          failureCode: PROVIDER_TEMPORARY_FAILURE
        createdAt: 2026-07-13T09:42:00Z
        updatedAt: 2026-07-13T09:42:01Z